requirements.txt
Pillow>=9.3.0
PyQRCode>=1.2.1
pypng>=0.20220715.0